perf: 优化ffi数据定义

+++ all/render_sort.c
@@ -3,8 +3,8 @@
 
 typedef struct
 {
-    double sort_y;
-    double pos_x;
+    float sort_y;
+    float pos_x;
     int    z;
     int    draw_order;
     int    lua_index;
